+++ WebKit/Source/WebCore/page/scrolling/ScrollingTree.h
@@ -59,7 +59,7 @@ class ScrollingTreeOverflowScrollProxyNode;
 class ScrollingTreePositionedNode;
 class ScrollingTreeScrollingNode;
 class ScrollingTreeFrameHostingNode;
-enum class EventListenerRegionType : uint32_t;
+enum class EventListenerRegionType : uint64_t;
 
 using FramesPerSecond = unsigned;
 using PlatformDisplayID = uint32_t;
